starter-kits/2-reasoning-agents/requirements.txt
# =============================================================================
# requirements.txt — dependencias de PRODUCCIÓN
# Para desarrollo/testing: pip install -r requirements-dev.txt
# =============================================================================

# --- Azure ---
azure-ai-projects>=1.0.0b11      # scripts de utilidad (list_foundry_agents.py)
azure-identity>=1.19.0            # credenciales Azure (requerido por azure-monitor)
azure-monitor-opentelemetry>=1.6.0  # telemetría → Application Insights

# --- AI / OpenAI ---
openai>=1.82.0

# --- Validación y configuración ---
pydantic>=2.11.0
pydantic-settings>=2.9.0
python-dotenv>=1.1.0

# --- Dashboard Web (Flask + SocketIO) ---
flask>=3.0.0
flask-socketio>=5.3.6
python-socketio>=5.11.0
eventlet>=0.38.0                  # worker async para gunicorn + SocketIO

# --- HTTP asíncrono ---
httpx>=0.27.0

# --- Servidor de producción ---
gunicorn>=23.0.0
